Move bwapi4 branch to main trunk.
git-svn-id: https://bwapi.googlecode.com/svn/trunk@4386 273ecf2c-f24b-0410-b6a7-e9298bb4873f

+++ bwapi/BWAPI/Source/DLLMain.h
@@ -1,9 +1,8 @@
 #pragma once
-#include <windows.h>
 #include <string>
 #include "BW/Offsets.h"
 
-void __fastcall QueueGameCommand(void *pBuffer, DWORD dwLength);
+void __fastcall QueueGameCommand(void *pBuffer, size_t dwLength);
 
 #define QUEUE_COMMAND(x,...) QueueGameCommand(&x(__VA_ARGS__), sizeof(x))
